Create a bucket for transition WAL-E backups
This is based on the infra-wal-e-warehouse-bucket project.
Only needed in production and staging as integration is using RDS.

This used to be configured in govuk-terraform-provisioning but
we can't deploy that now, so I've created new buckets here and my
plan is to configure WAL-E to use the new ones and copy the data
across from the old ones.